+++ src/middleware/guest.ts
@@ -1,11 +1,11 @@
-import {MiddlewareParams} from '../../types';
+import {MiddlewareParams} from '../types';
 
 export const guestMiddleware = ({
   loggedIn,
   next,
   options,
 }: MiddlewareParams) => {
-  if (loggedIn.value) {
+  if (loggedIn?.value) {
     next(options.redirect.home);
   } else {
     next();