+++ src/storage/drivers.ts
@@ -1,7 +1,7 @@
 import {CookieStorage} from './cookie';
 import {LocalStorage} from './local';
 import {SecureLocalStorage} from './secure-ls';
-import {AuthStorage, SupportedAuthStorage} from '../../types/index';
+import {AuthStorage, SupportedAuthStorage} from '../types/index';
 
 export const drivers = {
   local: LocalStorage,