+++ CHANGELOG.md
@@ -2,6 +2,83 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.6](https://github.com/pinteraktif/smart-akmil-web/compare/v0.1.5...v0.1.6) (2021-11-12)
+
+
+### Features
+
+* allow to send extra payload to logout method ([c875622](https://github.com/pinteraktif/smart-akmil-web/commit/c875622bbccccc275ab6471e386090b863706a54))
+
+
+### Bug Fixes
+
+* type path ([0083f54](https://github.com/pinteraktif/smart-akmil-web/commit/0083f545e37b344eca928e04ffd312d41f687846))
+
+### [0.1.5](https://git.gits.id/frontend/starter/vue-3/starter/compare/v0.1.4...v0.1.5) (2021-11-05)
+
+
+### Bug Fixes
+
+* require lodash ([19227a8](https://git.gits.id/frontend/starter/vue-3/starter/commit/19227a8ec6fffea6a7bada9a583bafd8f5f9d87e))
+
+### [0.1.4](https://git.gits.id/frontend/starter/vue-3/starter/compare/v0.1.3...v0.1.4) (2021-11-05)
+
+
+### Bug Fixes
+
+* disable refresh token when error from login page ([cb8af87](https://git.gits.id/frontend/starter/vue-3/starter/commit/cb8af8762f4983cb86e4a057ae8421552c09261f))
+* force logout user when failed to refresh token ([5b268d2](https://git.gits.id/frontend/starter/vue-3/starter/commit/5b268d2d586ca721f7e6bd9a6dd67d05d8362c18))
+
+### [0.1.3](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/compare/v0.1.2...v0.1.3) (2021-11-05)
+
+
+### Bug Fixes
+
+* rework refresh token flow ([5ea1cd2](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/commit/5ea1cd27959a44cb920236c6b80ae8afcfabfea5))
+
+### [0.1.2](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/compare/v0.1.1...v0.1.2) (2021-11-05)
+
+
+### Bug Fixes
+
+* uses index.ts as main entry point for now ([475c820](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/commit/475c820036b01e239db851bf2c09dacf554db8f8))
+
+### [0.1.1](https://git.gits.id/frontend/starter/vue-3/starter/compare/v0.1.0...v0.1.1) (2021-11-05)
+
+
+### Features
+
+* export src/tyles ([ea8588c](https://git.gits.id/frontend/starter/vue-3/starter/commit/ea8588c411b1510c4c0bb9138765feef4168c640))
+
+## [0.1.0](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/compare/v0.0.15...v0.1.0) (2021-11-05)
+
+
+### Features
+
+* improved typing ([d7ea3ed](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/commit/d7ea3ed643f2b3cbb4d3b08b3135568fd62ac4bd))
+
+### [0.0.15](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/compare/v0.0.14...v0.0.15) (2021-11-05)
+
+
+### Bug Fixes
+
+* plugin registration ([869c754](https://gitlab.com/wartek-id/belajar-id/belajar.id-cms/commit/869c754c3edb2f51d1ef37a3772e5b4f43815a51))
+
+### [0.0.14](https://github.com/pinteraktif/smart-akmil-web/compare/v0.0.13...v0.0.14) (2021-11-04)
+
+
+### Bug Fixes
+
+* clear refresh token storage ([94b6547](https://github.com/pinteraktif/smart-akmil-web/commit/94b654745c621a1c85367a351a074cb755973856))
+
+### [0.0.13](https://github.com/pinteraktif/smart-akmil-web/compare/v0.0.12...v0.0.13) (2021-11-04)
+
+
+### Bug Fixes
+
+* exp date calculation ([713cf55](https://github.com/pinteraktif/smart-akmil-web/commit/713cf5563bf3ba7e7041a2b60041affef80519ad))
+* refresh token flow and maxAge ([f106fd2](https://github.com/pinteraktif/smart-akmil-web/commit/f106fd2982424abd56e5f7e3c290d9fbadb870a2))
+
 ### [0.0.12](https://git.gits.id/frontend/starter/vue-3/starter/compare/v0.0.11...v0.0.12) (2021-11-04)